requirements.txt
pandas
numpy
faker
scikit-learn
langchain
langchain-openai
openai
sqlalchemy
plotly
pyyaml 
langchain-community
